refactor(dropdown.tsx): move dropdown component to global components folder
